refactor(cli): preserve parser contract facade
