+++ packages/cli/src/cli/thin-command.ts
@@ -16,7 +16,6 @@ export { firstCliCommandIndex } from "./thin-command-help.ts";
 export { firstCliCommand } from "./thin-command-help.ts";
 export { helpDomain } from "./thin-command-help.ts";
 export { deriveThinCliInputs } from "./thin-command-inputs.ts";
-export { parseThinCommand } from "./thin-command-router.ts";
 export { cliCommandDomains } from "./thin-command-help.ts";
 export { unsupportedCommandHint } from "./thin-command-help.ts";
 export type { ThinHelpCatalogEntry } from "./thin-command-help.ts";
@@ -24,9 +23,89 @@ export type { ThinCommand } from "./thin-command-types.ts";
 export type { ThinParseResult } from "./thin-command-types.ts";
 export type { ThinCliInput } from "./thin-command-types.ts";
 
-import { thinCliCommands } from "../../../daemon/src/protocol/daemon-protocol.contract.ts";
-import { commandDomains } from "./thin-command-help.ts";
+import {
+  daemonProtocolCommands,
+  resolveThinCliCommand,
+  safePath,
+  thinCliCommands,
+} from "../../../daemon/src/protocol/daemon-protocol.contract.ts";
+import type { SafePath } from "../../../daemon/src/protocol/daemon-protocol.contract.ts";
+import {
+  accepted,
+  globalOption,
+  nonEmpty,
+  rejected,
+  stripGlobals,
+} from "./thin-command-flags.ts";
+import {
+  commandDomains,
+  unsupportedCommandHint,
+} from "./thin-command-help.ts";
 import type { ThinHelpCatalogEntry } from "./thin-command-help.ts";
+import { deriveInputDirectory } from "./thin-command-inputs.ts";
+import { parseRouted } from "./thin-command-router.ts";
+import { parseResumeDispatch } from "./thin-command-runtime.ts";
+import { parseTask } from "./thin-command-task.ts";
+import type {
+  ProtocolCommand,
+  ThinCliInputDirectory,
+  ThinParseResult,
+} from "./thin-command-types.ts";
+
+type LifecyclePublicationAction =
+  | "task-create"
+  | "task-start"
+  | "task-submit"
+  | "task-review-execution"
+  | "task-complete";
+
+export function parseThinCommand(
+  argv: readonly string[],
+  cwd = process.cwd(),
+  commands: readonly ProtocolCommand[] = daemonProtocolCommands,
+): ThinParseResult {
+  const rootDir = safePath(globalOption(argv, "--root") ?? cwd),
+    repoId = globalOption(argv, "--repo"),
+    json = argv.includes("--json"),
+    args = stripGlobals(argv),
+    route =
+      commands === daemonProtocolCommands
+        ? resolveThinCliCommand(args)
+        : commands.find((entry) =>
+            entry.path.every((token, index) => args[index] === token),
+          ),
+    inputs = deriveInputDirectory(route);
+  if (
+    route?.id === "runtime-run" &&
+    args[2]?.startsWith("--") &&
+    args.includes("--resume-dispatch")
+  )
+    return parseResumeDispatch(rootDir, repoId, json, args, inputs);
+  if (route?.id === "task-dispatches" && nonEmpty(args[2]) && args.length === 3)
+    return accepted(
+      rootDir,
+      repoId,
+      json,
+      { kind: "task-dispatches", taskId: args[2] },
+      "repo.task.dispatches",
+    );
+  const routed = parseRouted(route, args, rootDir, repoId, json, inputs);
+  if (routed) return routed;
+  if (!route || args[0] !== "task")
+    return rejected("unsupported_command", unsupportedCommandHint(args), json);
+  return parseTaskRoute(route.id, args, rootDir, repoId, json, inputs);
+}
+
+function parseTaskRoute(
+  id: ProtocolCommand["id"] | LifecyclePublicationAction,
+  args: readonly string[],
+  rootDir: SafePath,
+  repoId: string | undefined,
+  json: boolean,
+  inputs: ThinCliInputDirectory,
+): ThinParseResult {
+  return parseTask(id, args, rootDir, repoId, json, inputs);
+}
 
 export function renderThinHelp(
   catalog: readonly ThinHelpCatalogEntry[] = [],
